Who's That Pokemon Finalised
After long gruelling weeks this is the finished product of Who's That Pokemon. Everything has been finalised and refurnished and it runs just as I had hoped. The code is programmed to spit out a Pokemon name from the original 151 and you have to guess the number the Pokemon is in the Pokedex.